requirements.txt
PySide6>=6.6.1
pyheif>=0.8.0
Pillow>=10.2.0
opencv-python>=4.9.0.80
SQLAlchemy>=2.0.25
Send2Trash>=1.8.2